change config for app
