+++ app/src/modules/view/requests.js
@@ -1,0 +1,29 @@
+import './view';
+
+const sendHttpRequest = (method, url, data) => {
+    console.log(JSON.stringify(data));
+    return fetch(url, {
+        method: method,
+        cache: 'reload',
+        body: JSON.stringify(data),
+        headers: data ? { 'Content-Type': 'application/json' } : {}
+    }).then(response => {
+        if (response.status >= 400) {
+            return response.json().then(errResData => {
+                const error = new Error('Something went wrong!');
+                error.data = errResData;
+                throw error;
+            });
+        }
+        return response.json();
+    });
+};
+
+const getData = () => {
+    sendHttpRequest('GET', 'http://localhost:5000/employees/list').then(responseData => {
+        for (let i = 0; i < responseData.length; i++) {
+            renderView(responseData[i]);
+        }
+        listenButtonsView();
+    });
+};
