+++ app/src/pages/index/index.pug
@@ -9,7 +9,4 @@ html(lang="en")
         body
             block header
             main
-                button.list='list'
-                button.add='add'
-                button.delete='delete'
                 include ../../modules/view/view
