+++ app/src/theme/colors.scss
@@ -1,0 +1,7 @@
+$dark_shade: #1f2041;
+$dark_shade_25: rgba(31, 32, 65, 0.25);
+$dark_shade_50: rgba(31, 32, 65, 0.5);
+$dark_shade_75: rgba(31, 32, 65, 0.75);
+$purple: #bc9cff;
+$green: #6fcf97;
+$gradient: linear-gradient(180deg, $purple, #8BA4F9);
